✨ Rollup config
Now correctly compiles the plugin, the import should work as expected.

Since we now use Rollup, three files are available:
A UMD version that can be used in a script import and will instantiate the Toasteo class and set it to window.Toasteo.
A CommonJS (for Node) and an ES Module (for bundlers) are also available.

The docs have been updated to reflect the changes.